+++ src/test/java/org/itsallcode/openfasttrace/gradle/OpenFastTracePluginTest.java
@@ -121,7 +121,7 @@ class OpenFastTracePluginTest
                             <specobject>
                               <id>exampleB</id>
                               <shortdesc>Tracing Example</shortdesc>
-                              <status>approved</status>
+                              <status>draft</status>
                               <version>1</version>
                         """,
 
@@ -186,7 +186,7 @@ class OpenFastTracePluginTest
         assertThat(buildResult.task(":traceRequirements").getOutcome(),
                 either(is(TaskOutcome.SUCCESS)).or(is(TaskOutcome.FROM_CACHE)));
         TestUtil.assertFileContent(PROJECT_CUSTOM_CONFIG_DIR.resolve("build/custom-report.txt"),
-                "not ok [ in:  1 /  1 ✔ | out:  0 /  0   ] dsn~exampleB~1 (impl, -utest)",
+                "not ok [ in:  1 /  1 ✔ | out:  0 /  0   ] dsn~exampleB~1 [draft] (impl, -utest)",
                 "not ok - 2 total, 1 direct, 0 transitive defects");
     }
 
@@ -198,7 +198,7 @@ class OpenFastTracePluginTest
         assertEquals(TaskOutcome.FAILED,
                 buildResult.task(":traceRequirements").getOutcome());
         TestUtil.assertFileContent(PROJECT_CUSTOM_CONFIG_DIR.resolve("build/custom-report.txt"),
-                "not ok [ in:  1 /  1 ✔ | out:  0 /  0   ] dsn~exampleB~1 (impl, -utest)",
+                "not ok [ in:  1 /  1 ✔ | out:  0 /  0   ] dsn~exampleB~1 [draft] (impl, -utest)",
                 "not ok - 2 total, 1 direct, 0 transitive defects");
     }
 
@@ -209,6 +209,33 @@ class OpenFastTracePluginTest
                 "traceRequirements", "-PfailBuild=true", "-PfilteredArtifactTypes=dsn");
         assertThat(buildResult.task(":traceRequirements").getOutcome(),
                 either(is(TaskOutcome.SUCCESS)).or(is(TaskOutcome.FROM_CACHE)));
+    }
+
+    @Test
+    void filteredWantedStatuses()
+    {
+        final BuildResult buildResult = runBuild(PROJECT_CUSTOM_CONFIG_DIR, "clean",
+                "traceRequirements",
+                "-PfilterWantedStatuses=draft,approved");
+        assertThat(buildResult.task(":traceRequirements").getOutcome(),
+                either(is(TaskOutcome.SUCCESS)).or(is(TaskOutcome.FROM_CACHE)));
+        TestUtil.assertFileContent(PROJECT_CUSTOM_CONFIG_DIR.resolve("build/custom-report.txt"),
+                "not ok [ in:  1 /  1 ✔ | out:  0 /  0   ] dsn~exampleB~1 [draft] (impl, -utest)",
+                "not ok - 2 total, 1 direct, 0 transitive defects");
+    }
+
+    @Test
+    void filteredWantedStatusesNoMatch()
+    {
+        final BuildResult buildResult = runBuild(PROJECT_CUSTOM_CONFIG_DIR, "clean",
+                "traceRequirements",
+                "-PfilterWantedStatuses=approved");
+        assertThat(buildResult.task(":traceRequirements").getOutcome(),
+                either(is(TaskOutcome.SUCCESS)).or(is(TaskOutcome.FROM_CACHE)));
+        TestUtil.assertFileContent(
+                PROJECT_CUSTOM_CONFIG_DIR.resolve("build/custom-report.txt"),
+                "not ok [ in:  0 /  0   | out:  0 /  1 ✘ ] impl~exampleB-34853351~0 ()",
+                "not ok - 1 total, 1 direct, 0 transitive defects");
     }
 
     @Test
@@ -257,7 +284,8 @@ class OpenFastTracePluginTest
         buildResult = runBuild(DEPENDENCY_CONFIG_DIR, "traceRequirements");
         assertThat(buildResult.task(":traceRequirements").getOutcome(),
                 either(is(TaskOutcome.SUCCESS)).or(is(TaskOutcome.FROM_CACHE)));
-        TestUtil.assertFileContent(DEPENDENCY_CONFIG_DIR.resolve("build/reports/tracing.txt"),
+        TestUtil.assertFileContent(
+                DEPENDENCY_CONFIG_DIR.resolve("build/reports/tracing.txt"),
                 "requirements-1.0.zip!spec.md:2",
                 "requirements-1.0.zip!source.java:1",
                 "not ok - 2 total, 1 direct, 0 transitive defects");
@@ -333,7 +361,8 @@ class OpenFastTracePluginTest
         }
         catch (final IOException e)
         {
-            throw new UncheckedIOException("Failed to create dependency zip " + dependencyZip, e);
+            throw new UncheckedIOException(
+                    "Failed to create dependency zip " + dependencyZip, e);
         }
     }